requirements.txt
django
graphene_django
django-graphql-playground
pytz
typing_extensions
django-graphql-jwt